Adapted tracking assignment
I adapted the tracking assignment to the contest. It uses an adapted
JointParticleFilter, but allows for elapsing the time of a single ghost
as opposed to all of them. This is useful since the each team takes
alternating turns. Having a shared particle filter is also especially
useful, I think, since it allows for the measurements of all of our
agents to combined.

I also added a strategy module. Not sure if this is the best way to go,
but it basically allows us to do sort of inversion of control on the
agents behavior function (we can change the strategy of an agent). This
is in contrast to just writing a bunch of agents. It may also allow for
a single strategy to coordinate multiple agents in a convinient manner.