Implemented feature based strategies
I implemented strategy.Feature, which offers the boilerplate code for strategies
to use features to control agents. I also implemented some basic features and a
very ineffective Offensive strategy. This commit is more about implementing the
Feature class than the offensive agent.

Feature based agents specify a dictionary of features extracted from a game
state as well as a dictionary of weights. The action which results in the
highest evaluation, the dot product of features and weights, is chosen. This is
very similar to the Q-learning assignment, and it seems that the weights can be
learned in a similar fashion.