Substantially improved tracking
The old tracking code was modifying the opponents' state objects in
gameState.agentStates, and as a result the positions of agents were not
correctly returning None when they were to far away to be observed. However,
more importantly, the code could not determine when the agents were close enough
to be observed exactly.

I changed this by copying the game state before calling elapseTime, and now
checking the positions inside of observeState. Now, if an agent is close enough
to be observed exactly, we update the particle filter accordingly. Now,
opponents can be tracked with much greater accuracy.