Added Strategic Ghosts
I implemented strategic ghosts, that is ghosts that behave according to a
strategy. Currently, it assumes both ghosts are BaselineOffensive, which
simulates one ghost perfectly, now that particles are initialized to the
starting position. However, the defensive ghost is improperly simulated.
Furthermore, this ghost is substantially slower than the random ghost, which
means that after reweighting particles, it takes a noticeably longer time to
complete a turn. At this point, the BaselineOffensive agent is fast enough so
that we don't overrun our turn. However, our current offensive agent takes to
long to simulate for the other agent. We can try to optimize, but it seems that
there is a hard upperbound on the amount of simulation logic we have for ghosts.

Also, in order to help aliviate some of the problem, I am going to try to change
the particle filter to allow for reweighting a single ghost, instead of both.
That way, we don't lose accurate observations of one ghost if we are poorly
simulating the other.