Implemented adaptive ghost agents
I have implemented adpative ghost agents. Basically, based on the amount of time
a given agent spends on either side of the board, it changes the way we simulate
it. If it spends more time on our side, we assume it is offensive. If we cannot
determine which side it spends the most time on, we simulate it randomly.

I have one qualm with the particle filter. Very frequently, it resets all the
particles because one ghost has drifted to far from where we believe it to be.
This has the unfortunate affect of clobbering the other ghost. This is
unfortunate because, for example, one ghost's belief cloud may be accurate,
while the other's is inaccurate. It also has the affect of making the
computation times much slower.