Refactored ghost code into strategy
In the previous commit, I implemented an adaptive ghost that, based on some
features, calculated which strategy to use to simulate the ghost. I refactored
this code out into a strategy. Aditionally, I introduced an Adaptive abstract
class that is a refinement of the Feature class that attempts to classify the
current board state into one of two optimal strategies. This could be useful for
future agents that can alternate between offensive and defensive based on score.