Cleaned up features and factory
I cleaned up some of the features. They should all be semantically the same, but
now I think they are clearer, and many don't use intermediate lists, as they did
before. The factory class also adapts to the absence of provided weights, in
which case it defaults to the weights given inside the respective classes.