Implemented perceptron learning; changed -w
I have implemented perceptron based learning with the 'learn' option. It takes a
list of agents to learn. This list must include indices on the current team. If
this option is passed, the agent runs a perceptron in addition to making its
regular moves. This is advantageous because, for example, we can write agents
that don't use weights and learn weights that approximate them.

I have also refactored the '-w' option to just use the team options instead.
This is nice, because it defaults to the config file, which will allow a way to
better organize things like default weights. We still need to change evolve.go
in order to account for this.