Implemented Negamax (without pruning)
I implemented the negamax algorithm, which is equivalent to minimax but slightly
simpler to implement for alternating adversary, friendly prediction. It seems to
work, although it may need more testing. It takes another strategy as input for
use as the heuristic. It traverses down the state tree to a certain depth and
applies the nested strategy, and then computes the appropriate min/max
functions.

Ideally, we can combine this function with the learning agent in order to learn
weights that approximate a better agent. This would, in some sense, allow us to
boot strap our agents weights to become increasingly better. It remains to be
seen if this will actually work.