Fixed learning agents; Alpha-beta pruning for Negamax
Firstly, I implemented alpha beta pruning for the negamax algorithm, which
should now execute faster on average. It is still probably to risky to run it at
4 for the contest, but a depth of 2 fits well within the time limit.

I also fixed the learning agents and keyboard agents, so now it seems that they
work well together. I also refactored the parameters to be first|second=<opts>.
The opts can be any of the following:
    strategy = ContestOffensive|ContestDefensive|Keyboard
    learn = ContestOffensive|ContestDefensive| any x in strategy.py

I still need to make it possible to seed the learning environment with weights,
and make it possible to learn from saved games. The former should not be to
difficult, but the latter may prove more challenging.