Can specify seed weights for perceptron
We can now specify the weights to initialize the perceptron using the weights
option to the agent. For example: --redOpts
{'first':{'learn':'ContestOffensive'|{'score':1.0}}} says to learn weights for
the contest offensive strategy starting with the weights {score: 1.0}. To
continue learning, we should thread the resulting weights from one round into
the seed weights for the next round.

I still have to implement learning from recorded games.