Learning from replay almost works
At this point, learning from a replay almost works. The only problem is that the
starting team is randomly selected inside of capture, so if the agent selected
to go first is different than the agent that was selected in the previous game,
then the game willl likely fail with an illegal move at some point.
Additionally, the map must be synchronized with the one that the game was run
on.

I am going to incorporate the changes into capture.py and the --replay option,
which at this point is totally unrelated to the changes I have made. However, if
the correct starting team is chosen, then the game will execute correctly.